+++ src/Http/Middleware/SetAuthDefaults.php
@@ -19,12 +19,12 @@ class SetAuthDefaults
      */
     public function handle($request, Closure $next)
     {
-        //dd($request->guard());
         if ($guard = $request->guard()) {
             // It's better to set auth defaults config globally,
             // instead of using `auth()->shouldUse($guard);`
             config()->set('auth.defaults.guard', $guard);
-            config()->set('auth.defaults.provider', Str::plural($guard));
+            config()->set('auth.defaults.apiguard', $guard);
+            config()->set('auth.defaults.provider', Str::afterLast(Str::plural($guard), ':'));
             config()->set('auth.defaults.passwords', $guard);
             config()->set('auth.defaults.emails', $guard);
         }